#1068  firebase에서 불러온 record를 마이코스 recyclerview에 보이도록 변경
- mockRecords를 recordsItems로 변경
