+++ app/src/main/java/smu/ac/kr/johnber/my/MyStatisticsPagerAdapter.java
@@ -39,13 +39,12 @@ public class MyStatisticsPagerAdapter extends PagerAdapter {
    *
    * @return
    */
-  private HashMap<String, Record> recordHashMap = new HashMap<String, Record>();
+
 
   @Override
   public int getCount() {
     return 3;
   }
-  HashMap<String, Record> mapRecord = new HashMap<>();
 
   @Override
   public Object instantiateItem(ViewGroup container, int position) {
@@ -57,7 +56,6 @@ public class MyStatisticsPagerAdapter extends PagerAdapter {
     switch (position) {
       case 0:
         viewId = R.layout.my_statistics_daily_view;
-        getRecord();
         break;
       case 1:
         viewId = R.layout.my_statistics_weekly_view;
@@ -73,42 +71,7 @@ public class MyStatisticsPagerAdapter extends PagerAdapter {
 
   }
 
-  public void getRecord(){
 
-
-    FirebaseDatabase database = FirebaseDatabase.getInstance();
-    DatabaseReference myRef = database.getReference();
-
-    myRef.addValueEventListener(new ValueEventListener() {
-      @Override
-      public void onDataChange(DataSnapshot dataSnapshot) {
-        FirebaseUser user = FirebaseAuth.getInstance().getCurrentUser();
-        String uid = user.getUid();
-
-        //DB에 저장된 데이터 HashMap에 저장.
-        for (DataSnapshot snapshot : dataSnapshot.getChildren()){
-          //DB에서 로그인한 아이디와 일치하는지 확인 후 해당 데이터만 읽어옴.
-          if (snapshot.getKey().toString().equals(uid)){
-            Log.d("MainActivity", "Single ValueEventListener : " + snapshot.getValue());
-            String keyDate1 = "";
-            for (DataSnapshot snapshot1 : snapshot.getChildren()){
-              keyDate1 = snapshot1.getKey().toString();
-              for (DataSnapshot snapshot2 : snapshot1.getChildren()){
-                String keyDate = keyDate1 + '/' + snapshot2.getKey().toString();
-                recordHashMap.put(keyDate, snapshot2.getValue(Record.class));
-                Log.d("mainactivity", "hashMap"+recordHashMap.toString());
-              }
-            }
-          }
-        }
-      }
-
-      @Override
-      public void onCancelled(DatabaseError databaseError) {
-
-      }
-    });
-  }
 
 
   @Override
